Refactor ColumnValueReader impl
