+++ kernel/data-pipeline/api/src/main/java/org/apache/shardingsphere/data/pipeline/spi/ingest/dumper/ColumnValueReader.java
@@ -18,6 +18,7 @@
 package org.apache.shardingsphere.data.pipeline.spi.ingest.dumper;
 
 import org.apache.shardingsphere.infra.util.spi.annotation.SingletonSPI;
+import org.apache.shardingsphere.infra.util.spi.type.required.RequiredSPI;
 import org.apache.shardingsphere.infra.util.spi.type.typed.TypedSPI;
 
 import java.sql.ResultSet;
@@ -28,7 +29,7 @@ import java.sql.SQLException;
  * Column value reader.
  */
 @SingletonSPI
-public interface ColumnValueReader extends TypedSPI {
+public interface ColumnValueReader extends TypedSPI, RequiredSPI {
     
     /**
      * Read column value.
