+++ test/pipeline/src/test/java/org/apache/shardingsphere/data/pipeline/core/ingest/dumper/ColumnValueReaderFactoryTest.java
@@ -34,8 +34,8 @@ public final class ColumnValueReaderFactoryTest {
     @Test
     public void assertGetInstance() {
         Collection<Pair<String, Class<? extends ColumnValueReader>>> paramResult = Arrays.asList(
-                Pair.of("MySQL", MySQLColumnValueReader.class),
-                Pair.of("PostgreSQL", PostgreSQLColumnValueReader.class), Pair.of("openGauss", PostgreSQLColumnValueReader.class));
+                Pair.of("PostgreSQL", PostgreSQLColumnValueReader.class), Pair.of("openGauss", PostgreSQLColumnValueReader.class),
+                Pair.of("MySQL", MySQLColumnValueReader.class), Pair.of("Oracle", DefaultColumnValueReader.class));
         for (Pair<String, Class<? extends ColumnValueReader>> each : paramResult) {
             ColumnValueReader actual = ColumnValueReaderFactory.getInstance(each.getKey());
             assertThat(actual, instanceOf(each.getValue()));
